+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 ## v1.3.1 - ...
 
 
-- ...
+- On first install or updates, if custom sentences have been updated,
+  tries to reload Home Assistant Assist sentences without asking for a new restart.  
+  Only if fails to reload, it will ask the user to restart.
 
 
 ## v1.3.0 - 29/01/2026